# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0",
        "title": "Twonana Portfolio Go",
        "contact": {}
    },
    "host": "127.0.0.1:3000",
    "basePath": "/",
    "securityDefinitions": {},
    "schemes": [
        "http"
    ],
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "paths": {
        "/api/v1/nft": {
            "post": {
                "summary": "Create NFT",
                "tags": [
                    "Nft"
                ],
                "operationId": "CreateNFT",
                "deprecated": false,
                "produces": [
                    "application/json"
                ],
                "consumes": [
                    "multipart/form-data"
                ],
                "parameters": [
                    {
                        "name": "title",
                        "in": "formData",
                        "required": true,
                        "type": "string",
                        "description": ""
                    },
                    {
                        "name": "image_url",
                        "in": "formData",
                        "required": true,
                        "type": "file",
                        "format": "file",
                        "description": ""
                    },
                    {
                        "name": "category",
                        "in": "formData",
                        "required": true,
                        "type": "string",
                        "description": ""
                    },
                    {
                        "name": "description",
                        "in": "formData",
                        "required": true,
                        "type": "string",
                        "description": ""
                    },
                    {
                        "name": "software",
                        "in": "formData",
                        "required": true,
                        "type": "string",
                        "description": ""
                    },
                    {
                        "name": "size",
                        "in": "formData",
                        "required": true,
                        "type": "string",
                        "description": ""
                    },
                    {
                        "name": "nft_format",
                        "in": "formData",
                        "required": true,
                        "type": "string",
                        "description": ""
                    },
                    {
                        "name": "marketplace_url",
                        "in": "formData",
                        "required": true,
                        "type": "string",
                        "description": ""
                    },
                    {
                        "name": "countdown_days",
                        "in": "formData",
                        "required": true,
                        "type": "integer",
                        "format": "int32",
                        "description": ""
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/CreateNFT"
                        },
                        "examples": {
                            "application/json": {
                                "meta": {
                                    "code": 201,
                                    "message": "Created"
                                },
                                "data": {
                                    "_id": "64ce2cc6bddba9def8818e37",
                                    "title": "baby don't hurt me",
                                    "image_url": "https://res.cloudinary.com/dypwg6tc4/image/upload/v1691233478/twonana%20nft/1e426798-0300-4bea-b2d3-58ade364fdac.png",
                                    "category": "dewa",
                                    "description": "blablabla",
                                    "software": "Adobe Illustrator",
                                    "size": "1000px x 1000px",
                                    "nft_format": "png",
                                    "marketplace_url": "url",
                                    "countdown_days": 2
                                }
                            }
                        },
                        "headers": {
                            "Date": {
                                "type": "string",
                                "default": "Sat, 05 Aug 2023 11:04:38 GMT"
                            },
                            "Content-Length": {
                                "type": "string",
                                "default": "405"
                            },
                            "Vary": {
                                "type": "string",
                                "default": "Origin"
                            },
                            "Access-Control-Allow-Origin": {
                                "type": "string",
                                "default": "*"
                            }
                        }
                    }
                }
            },
            "get": {
                "summary": "Get List NFT",
                "tags": [
                    "Nft"
                ],
                "operationId": "GetListNFT",
                "deprecated": false,
                "produces": [
                    "application/json"
                ],
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/GetListNFT"
                        },
                        "examples": {
                            "application/json": {
                                "meta": {
                                    "code": 200,
                                    "message": "OK"
                                },
                                "data": [
                                    {
                                        "_id": "64c8d83b0d878f31dc1467e4",
                                        "title": "ganjar",
                                        "image_url": "https://res.cloudinary.com/dypwg6tc4/image/upload/v1690884155/twonana%20nft/3e05368d-c8e1-410e-bc60-28c2a96d944f.jpg",
                                        "category": "PDIP",
                                        "description": "blablabla",
                                        "software": "Adobe Illustrator",
                                        "size": "1000px x 1000px",
                                        "nft_format": "png",
                                        "marketplace_url": "url",
                                        "countdown_days": 2
                                    },
                                    {
                                        "_id": "64c8d9840d878f31dc1467e5",
                                        "title": "baby don't hurt me",
                                        "image_url": "https://res.cloudinary.com/dypwg6tc4/image/upload/v1690884484/twonana%20nft/818447a9-84e9-4045-8e75-11cc2b72af30.png",
                                        "category": "dewa",
                                        "description": "blablabla",
                                        "software": "Adobe Illustrator",
                                        "size": "1000px x 1000px",
                                        "nft_format": "png",
                                        "marketplace_url": "url",
                                        "countdown_days": 2
                                    },
                                    {
                                        "_id": "64ce2cc6bddba9def8818e37",
                                        "title": "baby don't hurt me",
                                        "image_url": "https://res.cloudinary.com/dypwg6tc4/image/upload/v1691233478/twonana%20nft/1e426798-0300-4bea-b2d3-58ade364fdac.png",
                                        "category": "dewa",
                                        "description": "blablabla",
                                        "software": "Adobe Illustrator",
                                        "size": "1000px x 1000px",
                                        "nft_format": "png",
                                        "marketplace_url": "url",
                                        "countdown_days": 2
                                    }
                                ]
                            }
                        },
                        "headers": {
                            "Date": {
                                "type": "string",
                                "default": "Sat, 05 Aug 2023 11:04:55 GMT"
                            },
                            "Content-Length": {
                                "type": "string",
                                "default": "1104"
                            },
                            "Vary": {
                                "type": "string",
                                "default": "Origin"
                            },
                            "Access-Control-Allow-Origin": {
                                "type": "string",
                                "default": "*"
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/nft/64c8d83b0d878f31dc1467e4": {
            "get": {
                "summary": "Get NFT By id",
                "tags": [
                    "Nft"
                ],
                "operationId": "GetNFTByid",
                "deprecated": false,
                "produces": [
                    "application/json"
                ],
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/GetNFTByid"
                        },
                        "examples": {
                            "application/json": {
                                "meta": {
                                    "code": 200,
                                    "message": "OK"
                                },
                                "data": {
                                    "_id": "64c8d83b0d878f31dc1467e4",
                                    "title": "ganjar",
                                    "image_url": "https://res.cloudinary.com/dypwg6tc4/image/upload/v1690884155/twonana%20nft/3e05368d-c8e1-410e-bc60-28c2a96d944f.jpg",
                                    "category": "PDIP",
                                    "description": "blablabla",
                                    "software": "Adobe Illustrator",
                                    "size": "1000px x 1000px",
                                    "nft_format": "png",
                                    "marketplace_url": "url",
                                    "countdown_days": 2
                                }
                            }
                        },
                        "headers": {
                            "Date": {
                                "type": "string",
                                "default": "Sat, 05 Aug 2023 11:05:24 GMT"
                            },
                            "Content-Length": {
                                "type": "string",
                                "default": "388"
                            },
                            "Vary": {
                                "type": "string",
                                "default": "Origin"
                            },
                            "Access-Control-Allow-Origin": {
                                "type": "string",
                                "default": "*"
                            }
                        }
                    }
                }
            },
            "put": {
                "summary": "Update NFT",
                "tags": [
                    "Nft"
                ],
                "operationId": "UpdateNFT",
                "deprecated": false,
                "produces": [
                    "application/json"
                ],
                "consumes": [
                    "multipart/form-data"
                ],
                "parameters": [
                    {
                        "name": "title",
                        "in": "formData",
                        "required": true,
                        "type": "string",
                        "description": ""
                    },
                    {
                        "name": "image_url",
                        "in": "formData",
                        "required": true,
                        "type": "file",
                        "format": "file",
                        "description": ""
                    },
                    {
                        "name": "category",
                        "in": "formData",
                        "required": true,
                        "type": "string",
                        "description": ""
                    },
                    {
                        "name": "description",
                        "in": "formData",
                        "required": true,
                        "type": "string",
                        "description": ""
                    },
                    {
                        "name": "software",
                        "in": "formData",
                        "required": true,
                        "type": "string",
                        "description": ""
                    },
                    {
                        "name": "size",
                        "in": "formData",
                        "required": true,
                        "type": "string",
                        "description": ""
                    },
                    {
                        "name": "nft_format",
                        "in": "formData",
                        "required": true,
                        "type": "string",
                        "description": ""
                    },
                    {
                        "name": "marketplace_url",
                        "in": "formData",
                        "required": true,
                        "type": "string",
                        "description": ""
                    },
                    {
                        "name": "countdown_days",
                        "in": "formData",
                        "required": true,
                        "type": "integer",
                        "format": "int32",
                        "description": ""
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/UpdateNFT"
                        },
                        "examples": {
                            "application/json": {
                                "meta": {
                                    "code": 200,
                                    "message": "OK"
                                },
                                "data": {
                                    "_id": "64c8d83b0d878f31dc1467e4",
                                    "title": "ganjar",
                                    "image_url": "https://res.cloudinary.com/dypwg6tc4/image/upload/v1690884155/twonana%20nft/3e05368d-c8e1-410e-bc60-28c2a96d944f.jpg",
                                    "category": "PDIP",
                                    "description": "blablabla",
                                    "software": "Adobe Illustrator",
                                    "size": "1000px x 1000px",
                                    "nft_format": "png",
                                    "marketplace_url": "url",
                                    "countdown_days": 2
                                }
                            }
                        },
                        "headers": {
                            "Date": {
                                "type": "string",
                                "default": "Sat, 05 Aug 2023 11:05:32 GMT"
                            },
                            "Content-Length": {
                                "type": "string",
                                "default": "388"
                            },
                            "Vary": {
                                "type": "string",
                                "default": "Origin"
                            },
                            "Access-Control-Allow-Origin": {
                                "type": "string",
                                "default": "*"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "CreateNFT": {
            "title": "CreateNFT",
            "example": {
                "meta": {
                    "code": 201,
                    "message": "Created"
                },
                "data": {
                    "_id": "64ce2cc6bddba9def8818e37",
                    "title": "baby don't hurt me",
                    "image_url": "https://res.cloudinary.com/dypwg6tc4/image/upload/v1691233478/twonana%20nft/1e426798-0300-4bea-b2d3-58ade364fdac.png",
                    "category": "dewa",
                    "description": "blablabla",
                    "software": "Adobe Illustrator",
                    "size": "1000px x 1000px",
                    "nft_format": "png",
                    "marketplace_url": "url",
                    "countdown_days": 2
                }
            },
            "type": "object",
            "properties": {
                "meta": {
                    "$ref": "#/definitions/Meta"
                },
                "data": {
                    "$ref": "#/definitions/Data"
                }
            },
            "required": [
                "meta",
                "data"
            ]
        },
        "Meta": {
            "title": "Meta",
            "example": {
                "code": 201,
                "message": "Created"
            },
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "format": "int32"
                },
                "message": {
                    "type": "string"
                }
            },
            "required": [
                "code",
                "message"
            ]
        },
        "Data": {
            "title": "Data",
            "example": {
                "_id": "64ce2cc6bddba9def8818e37",
                "title": "baby don't hurt me",
                "image_url": "https://res.cloudinary.com/dypwg6tc4/image/upload/v1691233478/twonana%20nft/1e426798-0300-4bea-b2d3-58ade364fdac.png",
                "category": "dewa",
                "description": "blablabla",
                "software": "Adobe Illustrator",
                "size": "1000px x 1000px",
                "nft_format": "png",
                "marketplace_url": "url",
                "countdown_days": 2
            },
            "type": "object",
            "properties": {
                "_id": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "image_url": {
                    "type": "string"
                },
                "category": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "software": {
                    "type": "string"
                },
                "size": {
                    "type": "string"
                },
                "nft_format": {
                    "type": "string"
                },
                "marketplace_url": {
                    "type": "string"
                },
                "countdown_days": {
                    "type": "integer",
                    "format": "int32"
                }
            },
            "required": [
                "_id",
                "title",
                "image_url",
                "category",
                "description",
                "software",
                "size",
                "nft_format",
                "marketplace_url",
                "countdown_days"
            ]
        },
        "GetListNFT": {
            "title": "GetListNFT",
            "example": {
                "meta": {
                    "code": 200,
                    "message": "OK"
                },
                "data": [
                    {
                        "_id": "64c8d83b0d878f31dc1467e4",
                        "title": "ganjar",
                        "image_url": "https://res.cloudinary.com/dypwg6tc4/image/upload/v1690884155/twonana%20nft/3e05368d-c8e1-410e-bc60-28c2a96d944f.jpg",
                        "category": "PDIP",
                        "description": "blablabla",
                        "software": "Adobe Illustrator",
                        "size": "1000px x 1000px",
                        "nft_format": "png",
                        "marketplace_url": "url",
                        "countdown_days": 2
                    },
                    {
                        "_id": "64c8d9840d878f31dc1467e5",
                        "title": "baby don't hurt me",
                        "image_url": "https://res.cloudinary.com/dypwg6tc4/image/upload/v1690884484/twonana%20nft/818447a9-84e9-4045-8e75-11cc2b72af30.png",
                        "category": "dewa",
                        "description": "blablabla",
                        "software": "Adobe Illustrator",
                        "size": "1000px x 1000px",
                        "nft_format": "png",
                        "marketplace_url": "url",
                        "countdown_days": 2
                    },
                    {
                        "_id": "64ce2cc6bddba9def8818e37",
                        "title": "baby don't hurt me",
                        "image_url": "https://res.cloudinary.com/dypwg6tc4/image/upload/v1691233478/twonana%20nft/1e426798-0300-4bea-b2d3-58ade364fdac.png",
                        "category": "dewa",
                        "description": "blablabla",
                        "software": "Adobe Illustrator",
                        "size": "1000px x 1000px",
                        "nft_format": "png",
                        "marketplace_url": "url",
                        "countdown_days": 2
                    }
                ]
            },
            "type": "object",
            "properties": {
                "meta": {
                    "$ref": "#/definitions/Meta"
                },
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/Data"
                    }
                }
            },
            "required": [
                "meta",
                "data"
            ]
        },
...
        "GetNFTByid": {
            "title": "GetNFTByid",
            "example": {
                "meta": {
                    "code": 200,
                    "message": "OK"
                },
                "data": {
                    "_id": "64c8d83b0d878f31dc1467e4",
                    "title": "ganjar",
                    "image_url": "https://res.cloudinary.com/dypwg6tc4/image/upload/v1690884155/twonana%20nft/3e05368d-c8e1-410e-bc60-28c2a96d944f.jpg",
                    "category": "PDIP",
                    "description": "blablabla",
                    "software": "Adobe Illustrator",
                    "size": "1000px x 1000px",
                    "nft_format": "png",
                    "marketplace_url": "url",
                    "countdown_days": 2
                }
            },
            "type": "object",
            "properties": {
                "meta": {
                    "$ref": "#/definitions/Meta"
                },
                "data": {
                    "$ref": "#/definitions/Data"
                }
            },
            "required": [
                "meta",
                "data"
            ]
        },
        "UpdateNFT": {
            "title": "UpdateNFT",
            "example": {
                "meta": {
                    "code": 200,
                    "message": "OK"
                },
                "data": {
                    "_id": "64c8d83b0d878f31dc1467e4",
                    "title": "ganjar",
                    "image_url": "https://res.cloudinary.com/dypwg6tc4/image/upload/v1690884155/twonana%20nft/3e05368d-c8e1-410e-bc60-28c2a96d944f.jpg",
                    "category": "PDIP",
                    "description": "blablabla",
                    "software": "Adobe Illustrator",
                    "size": "1000px x 1000px",
                    "nft_format": "png",
                    "marketplace_url": "url",
                    "countdown_days": 2
                }
            },
            "type": "object",
            "properties": {
                "meta": {
                    "$ref": "#/definitions/Meta"
                },
                "data": {
                    "$ref": "#/definitions/Data"
                }
            },
            "required": [
                "meta",
                "data"
            ]
        },
...
    },
...
}